backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-multipart==0.0.20
websockets==14.1

